requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0

# Training and evaluation
tqdm>=4.65.0
matplotlib>=3.7.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# API server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Optional: Experiment tracking
# wandb>=0.15.0

# Jupyter notebooks
jupyter>=1.0.0
